# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler",
        "title": "🚀 Effective Mobile",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GitHub",
            "url": "https://github.com/Sanchir01"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/wallet": {
            "post": {
                "description": "send money",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "parameters": [
                    {
                        "description": "send money",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wallets.SendCoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallets.SendCoinResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/wallets": {
            "get": {
                "description": "get all wallets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallets.GetAllWalletsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/wallets/create": {
            "post": {
                "description": "get balance wallet by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "parameters": [
                    {
                        "description": "create wallet",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wallets.CreateWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallets.CreateWalletResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/wallets/{id}": {
            "get": {
                "description": "get balance wallet by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wallets.GetBalanceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "wallets.CreateWalletRequest": {
            "type": "object",
            "required": [
                "balance",
                "currency"
            ],
            "properties": {
                "balance": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
        "wallets.CreateWalletResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "string"
                },
                "response": {
                    "$ref": "#/definitions/api.Response"
                }
            }
        },
        "wallets.GetAllWalletsResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "$ref": "#/definitions/api.Response"
                },
                "wallets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wallets.Wallet"
                    }
                }
            }
        },
        "wallets.GetBalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "response": {
                    "$ref": "#/definitions/api.Response"
                }
            }
        },
        "wallets.OperationType": {
            "type": "string",
            "enum": [
                "DEPOSIT",
                "WITHDRAW",
                "TRANSFER"
            ],
            "x-enum-varnames": [
                "OperationTypeDeposit",
                "OperationTypeWithdraw",
                "OperationTypeTransfer"
            ]
        },
        "wallets.SendCoinRequest": {
            "type": "object",
            "required": [
                "amount",
                "sender_id",
                "type",
                "wallet_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "sender_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/wallets.OperationType"
                },
                "wallet_id": {
                    "type": "string"
                }
            }
        },
        "wallets.SendCoinResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "string"
                },
                "response": {
                    "$ref": "#/definitions/api.Response"
                }
            }
        },
        "wallets.Wallet": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}